Move some exported data types to the root module
