+++ src/parse.rs
@@ -31,6 +31,7 @@ use crate::linklabel::{scan_link_label_rest, LinkLabel, ReferenceLabel};
 use crate::scanners::*;
 use crate::strings::CowStr;
 use crate::tree::{Tree, TreeIndex};
+use crate::{Alignment, CodeBlockKind, Event, LinkType, Options, Tag};
 
 // Allowing arbitrary depth nested parentheses inside link destinations
 // can create denial of service vulnerabilities if we're not careful.
@@ -38,160 +39,6 @@ use crate::tree::{Tree, TreeIndex};
 // explicitly allowed by the spec as long as the limit is at least 3:
 // https://spec.commonmark.org/0.29/#link-destination
 const LINK_MAX_NESTED_PARENS: usize = 5;
-
-/// Codeblock kind.
-#[derive(Clone, Debug, PartialEq)]
-pub enum CodeBlockKind<'a> {
-    Indented,
-    /// The value contained in the tag describes the language of the code, which may be empty.
-    Fenced(CowStr<'a>),
-}
-
-impl<'a> CodeBlockKind<'a> {
-    pub fn is_indented(&self) -> bool {
-        match *self {
-            CodeBlockKind::Indented => true,
-            _ => false,
-        }
-    }
-
-    pub fn is_fenced(&self) -> bool {
-        match *self {
-            CodeBlockKind::Fenced(_) => true,
-            _ => false,
-        }
-    }
-}
-
-/// Tags for elements that can contain other elements.
-#[derive(Clone, Debug, PartialEq)]
-pub enum Tag<'a> {
-    /// A paragraph of text and other inline elements.
-    Paragraph,
-
-    /// A heading. The field indicates the level of the heading.
-    Heading(u32),
-
-    BlockQuote,
-    /// A code block.
-    CodeBlock(CodeBlockKind<'a>),
-
-    /// A list. If the list is ordered the field indicates the number of the first item.
-    /// Contains only list items.
-    List(Option<u64>), // TODO: add delim and tight for ast (not needed for html)
-    /// A list item.
-    Item,
-    /// A footnote definition. The value contained is the footnote's label by which it can
-    /// be referred to.
-    FootnoteDefinition(CowStr<'a>),
-
-    /// A table. Contains a vector describing the text-alignment for each of its columns.
-    Table(Vec<Alignment>),
-    /// A table header. Contains only `TableRow`s. Note that the table body starts immediately
-    /// after the closure of the `TableHead` tag. There is no `TableBody` tag.
-    TableHead,
-    /// A table row. Is used both for header rows as body rows. Contains only `TableCell`s.
-    TableRow,
-    TableCell,
-
-    // span-level tags
-    Emphasis,
-    Strong,
-    Strikethrough,
-
-    /// A link. The first field is the link type, the second the destination URL and the third is a title.
-    Link(LinkType, CowStr<'a>, CowStr<'a>),
-
-    /// An image. The first field is the link type, the second the destination URL and the third is a title.
-    Image(LinkType, CowStr<'a>, CowStr<'a>),
-}
-
-/// Type specifier for inline links. See [the Tag::Link](enum.Tag.html#variant.Link) for more information.
-#[derive(Clone, Debug, PartialEq, Copy)]
-pub enum LinkType {
-    /// Inline link like `[foo](bar)`
-    Inline,
-    /// Reference link like `[foo][bar]`
-    Reference,
-    /// Reference without destination in the document, but resolved by the broken_link_callback
-    ReferenceUnknown,
-    /// Collapsed link like `[foo][]`
-    Collapsed,
-    /// Collapsed link without destination in the document, but resolved by the broken_link_callback
-    CollapsedUnknown,
-    /// Shortcut link like `[foo]`
-    Shortcut,
-    /// Shortcut without destination in the document, but resolved by the broken_link_callback
-    ShortcutUnknown,
-    /// Autolink like `<http://foo.bar/baz>`
-    Autolink,
-    /// Email address in autolink like `<john@example.org>`
-    Email,
-}
-
-impl LinkType {
-    fn to_unknown(self) -> Self {
-        match self {
-            LinkType::Reference => LinkType::ReferenceUnknown,
-            LinkType::Collapsed => LinkType::CollapsedUnknown,
-            LinkType::Shortcut => LinkType::ShortcutUnknown,
-            _ => unreachable!(),
-        }
-    }
-}
-
-/// Markdown events that are generated in a preorder traversal of the document
-/// tree, with additional `End` events whenever all of an inner node's children
-/// have been visited.
-#[derive(Clone, Debug, PartialEq)]
-pub enum Event<'a> {
-    /// Start of a tagged element. Events that are yielded after this event
-    /// and before its corresponding `End` event are inside this element.
-    /// Start and end events are guaranteed to be balanced.
-    Start(Tag<'a>),
-    /// End of a tagged element.
-    End(Tag<'a>),
-    /// A text node.
-    Text(CowStr<'a>),
-    /// An inline code node.
-    Code(CowStr<'a>),
-    /// An HTML node.
-    Html(CowStr<'a>),
-    /// A reference to a footnote with given label, which may or may not be defined
-    /// by an event with a `Tag::FootnoteDefinition` tag. Definitions and references to them may
-    /// occur in any order.
-    FootnoteReference(CowStr<'a>),
-    /// A soft line break.
-    SoftBreak,
-    /// A hard line break.
-    HardBreak,
-    /// A horizontal ruler.
-    Rule,
-    /// A task list marker, rendered as a checkbox in HTML. Contains a true when it is checked.
-    TaskListMarker(bool),
-}
-
-/// Table column text alignment.
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum Alignment {
-    /// Default text alignment.
-    None,
-    Left,
-    Center,
-    Right,
-}
-
-bitflags! {
-    /// Option struct containing flags for enabling extra features
-    /// that are not part of the CommonMark spec.
-    pub struct Options: u32 {
-        const ENABLE_TABLES = 1 << 1;
-        const ENABLE_FOOTNOTES = 1 << 2;
-        const ENABLE_STRIKETHROUGH = 1 << 3;
-        const ENABLE_TASKLISTS = 1 << 4;
-        const ENABLE_SMART_PUNCTUATION = 1 << 5;
-    }
-}
 
 #[derive(Debug, Default, Clone, Copy)]
 pub(crate) struct Item {
